requirements.txt
streamlit
requests
pillow
streamlit-cropper